+++ WebJar.Application/ViewModel/PropertyValueViewModel.cs
@@ -1,5 +1,5 @@
 ﻿namespace WebJar.Application.ViewModel
 {
-    public record class PropertyValueViewModel(string Value, float Price,
+    public record class PropertyValueViewModel(string Value, decimal Price,
         string PriceType = "CONSTANT");
 }
